creating another directory that holds screens, and makes an home file as an entry point
